Replace fragment with composable resource screen
